+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 -----------
 
 - Ability to remove headers from email message. `#24`_
+- Improved attachments interface. `#18`_
 - Support for sending single email. `#11`_
 
 0.1.1 - 05.10.2016
@@ -21,4 +22,5 @@ Changelog
 
 
 .. _#24: https://github.com/FriendlyCoders/postmarker/issues/24
+.. _#18: https://github.com/FriendlyCoders/postmarker/issues/18
 .. _#11: https://github.com/FriendlyCoders/postmarker/issues/11